Split into a few packages and fix regression regarding display result of step
