+++ core/src/main/resources/org/jobrunr/dashboard/frontend/src/components/jobs/history-chart/gantt-components/gantt-row-duration.js
