+++ core/src/main/resources/org/jobrunr/dashboard/frontend/src/components/jobs/history-chart/gantt-components/gantt-timeline-entry.js
@@ -1,5 +1,5 @@
 import Tooltip from "@mui/material/Tooltip";
-import {SwitchableTimeFormatter} from "../../utils/time-ago.js";
+import {SwitchableTimeFormatter} from "../../../utils/time-ago.js";
 import Typography from "@mui/material/Typography";
 
 export const GanttTimelineEntry = ({time, position}) => {
