+++ core/src/main/resources/org/jobrunr/dashboard/frontend/src/components/jobs/history-chart/timeline-data-components/determine-skipped-steps.js
@@ -1,5 +1,5 @@
-import {comparePreciseDates, dateAsMilliseconds} from "../../../utils/helper-functions.js";
-import {RUN_STEP_ONCE} from "../../utils/state-names.js";
+import {comparePreciseDates, dateAsMilliseconds} from "../../../../utils/helper-functions.js";
+import {RUN_STEP_ONCE} from "../../../utils/state-names.js";
 
 const earliestEntryForRetry = (historyByStep, stepOrder, attemptId) =>
     stepOrder
