+++ core/src/main/resources/org/jobrunr/dashboard/frontend/src/components/jobs/history-chart/timeline-data-components/timeline-axis.js
@@ -1,5 +1,5 @@
-import {dateAsMilliseconds, formatDuration} from "../../../utils/helper-functions.js";
-import {END_STATES, PROCESSING, RUN_STEP_ONCE} from "../../utils/state-names.js";
+import {dateAsMilliseconds, formatDuration} from "../../../../utils/helper-functions.js";
+import {END_STATES, PROCESSING, RUN_STEP_ONCE} from "../../../utils/state-names.js";
 
 const COMPRESSOR_FLOOR_RATIO = 0.15;
 const TICK_MERGE_PCT = 8;
