+++ core/src/main/resources/org/jobrunr/dashboard/frontend/src/components/jobs/history-chart/timeline-gantt-chart.js
@@ -4,17 +4,17 @@ import Tooltip from '@mui/material/Tooltip';
 import {useTheme} from "@mui/material";
 import {Rhombus, RhombusOutline} from "mdi-material-ui";
 import {Fragment} from 'react';
-import {Legend} from "./legend.js";
+import {Legend} from "./gantt-components/legend.js";
 import {formatDuration} from "../../../utils/helper-functions.js";
-import {GanttBar, getBarColor} from "./gantt-bar.js";
-import {BreakIndicator} from "./break-indicator.js";
-import {RetrySeparator} from "./retry-separator.js";
+import {GanttBar, getBarColor} from "./gantt-components/gantt-bar.js";
+import {BreakIndicator} from "./gantt-components/break-indicator.js";
+import {RetrySeparator} from "./gantt-components/retry-separator.js";
 import {FAILED} from "../../utils/state-names.js";
 import {TIMELINE_MODES} from "./job-history-chart.js";
-import {GanttTooltipTitle} from "./gantt-tooltip-title.js";
-import {GanttRow} from "./gantt-row.js";
-import {GanttRowDuration} from "./gantt-row-duration.js";
-import {GanttTimelineEntry} from "./gantt-timeline-entry.js";
+import {GanttTooltipTitle} from "./gantt-components/gantt-tooltip-title.js";
+import {GanttRow} from "./gantt-components/gantt-row.js";
+import {GanttRowDuration} from "./gantt-components/gantt-row-duration.js";
+import {GanttTimelineEntry} from "./gantt-components/gantt-timeline-entry.js";
 
 export const MIN_LABEL_WIDTH = 150;
 export const MAX_LABEL_WIDTH = 250;
