+++ core/src/test/java/org/jobrunr/dashboard/FrontEndDevelopment.java
@@ -81,13 +81,13 @@ public class FrontEndDevelopment {
                         Map.entry("jr_step_first-step__6", true),
                         Map.entry("jr_step_start_first-step__6", now.minus(4 * 60 - 21, ChronoUnit.MINUTES).toString()),
                         Map.entry("jr_step_end_first-step__6", now.minus(4 * 60 - 23, ChronoUnit.MINUTES).toString()),
-                        Map.entry("jr_step_result_first-step__6", "Some amazing result"),
-                        Map.entry("jr_step_result_class_first-step__6", "java.lang.String"),
+                        Map.entry("jr_step_result_first-step", "Some amazing result"),
+                        Map.entry("jr_step_result_class_first-step", "java.lang.String"),
                         Map.entry("jr_step_second-step__6", true),
                         Map.entry("jr_step_start_second-step__6", now.minus(4 * 60 - 23, ChronoUnit.MINUTES).toString()),
                         Map.entry("jr_step_end_second-step__6", now.minus(3 * 60, ChronoUnit.MINUTES).toString()),
-                        Map.entry("jr_step_result_second-step__6", "Some second amazing result"),
-                        Map.entry("jr_step_result_class_second-step__6", "java.lang.String"),
+                        Map.entry("jr_step_result_second-step", "Some second amazing result"),
+                        Map.entry("jr_step_result_class_second-step", "java.lang.String"),
                         Map.entry("jr_step_third-step__6", false),
                         Map.entry("jr_step_start_third-step__6", now.minus(3 * 60, ChronoUnit.MINUTES).toString()),
                         Map.entry("jr_step_end_third-step__6", now.minus(2 * 60, ChronoUnit.MINUTES).toString()),
@@ -95,18 +95,18 @@ public class FrontEndDevelopment {
                         Map.entry("jr_step_third-step__10", true),
                         Map.entry("jr_step_start_third-step__10", now.minus(2 * 60 - 44, ChronoUnit.MINUTES).toString()),
                         Map.entry("jr_step_end_third-step__10", now.minus((2 * 60 - 44) * 60 - 10, ChronoUnit.SECONDS).toString()),
-                        Map.entry("jr_step_result_third-step__10", "Some third amazing result"),
-                        Map.entry("jr_step_result_class_third-step__10", "java.lang.String"),
+                        Map.entry("jr_step_result_third-step", "Some third amazing result"),
+                        Map.entry("jr_step_result_class_third-step", "java.lang.String"),
                         Map.entry("jr_step_fourth-step__10", true),
                         Map.entry("jr_step_start_fourth-step__10", now.minus(2 * 60 - 50, ChronoUnit.MINUTES).toString()),
                         Map.entry("jr_step_end_fourth-step__10", now.minus(60, ChronoUnit.MINUTES).toString()),
-                        Map.entry("jr_step_result_fourth-step__10", "Some fourth amazing result"),
-                        Map.entry("jr_step_result_class_fourth-step__10", "java.lang.String"),
+                        Map.entry("jr_step_result_fourth-step", "Some fourth amazing result"),
+                        Map.entry("jr_step_result_class_fourth-step", "java.lang.String"),
                         Map.entry("jr_step_fifth-step__10", true),
                         Map.entry("jr_step_start_fifth-step__10", now.minus(60, ChronoUnit.MINUTES).toString()),
                         Map.entry("jr_step_end_fifth-step__10", now.minus(10, ChronoUnit.SECONDS).toString()),
-                        Map.entry("jr_step_result_fifth-step__10", "Some fifth amazing result"),
-                        Map.entry("jr_step_result_class_fifth-step__10", "java.lang.String")
+                        Map.entry("jr_step_result_fifth-step", "Some fifth amazing result"),
+                        Map.entry("jr_step_result_class_fifth-step", "java.lang.String")
                 ))
                 .build());
